+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Contract breaking changes:
   * Database migration will handle a new structure for legacy data
   * The submit endpoint has been merged with the update endpoint
   * The GET all POD endpoint will filter out those PODs to which user has no right.
+* [OLMIS-3958](https://openlmis.atlassian.net/browse/OLMIS-3958): Ordered Quantity value is now in packs. Filled Quantity and Packs To Ship have been removed.
 
 New features:
 * [OLMIS-3663](https://openlmis.atlassian.net/browse/OLMIS-3663): Created Shipment resource model